Renamed Keyboard component to Shortcuts
