+++ src/components/Validation/Shortcuts.vue
